Water shader fixes
Added lighting to water shader + merged it into the toon shader to allow for smooth transitions btw water tiles and other tiles. The transition should be performed by changing the float parameter _IsWater (1f for water, 0f for other tiles) and setting the _Color (white for water) in the material's shader.